Moved PrivateRoute to hoc/ folder, <Router> to index.jsx, .gitkeep cleanup
