+++ public/index.html
@@ -1,53 +1,102 @@
 <!DOCTYPE html>
 <html lang="en">
+  <head>
+    <!-- Global site tag (gtag.js) - Google Analytics -->
+    <script
+      async
+      src="https://www.googletagmanager.com/gtag/js?id=UA-150018840-1"
+    ></script>
+    <script>
+      window.dataLayer = window.dataLayer || [];
 
-<head>
-  <!-- Global site tag (gtag.js) - Google Analytics -->
-  <script async src="https://www.googletagmanager.com/gtag/js?id=UA-150018840-1"></script>
-  <script>
-    window.dataLayer = window.dataLayer || [];
+      function gtag() {
+        dataLayer.push(arguments);
+      }
+      gtag("js", new Date());
 
-    function gtag() {
-      dataLayer.push(arguments);
-    }
-    gtag('js', new Date());
+      gtag("config", "UA-150018840-1");
+    </script>
 
-    gtag('config', 'UA-150018840-1');
-  </script>
+    <meta charset="utf-8" />
+    <link rel="apple-touch-icon" sizes="57x57" href="/apple-icon-57x57.png" />
+    <link rel="apple-touch-icon" sizes="60x60" href="/apple-icon-60x60.png" />
+    <link rel="apple-touch-icon" sizes="72x72" href="/apple-icon-72x72.png" />
+    <link rel="apple-touch-icon" sizes="76x76" href="/apple-icon-76x76.png" />
+    <link
+      rel="apple-touch-icon"
+      sizes="114x114"
+      href="/apple-icon-114x114.png"
+    />
+    <link
+      rel="apple-touch-icon"
+      sizes="120x120"
+      href="/apple-icon-120x120.png"
+    />
+    <link
+      rel="apple-touch-icon"
+      sizes="144x144"
+      href="/apple-icon-144x144.png"
+    />
+    <link
+      rel="apple-touch-icon"
+      sizes="152x152"
+      href="/apple-icon-152x152.png"
+    />
+    <link
+      rel="apple-touch-icon"
+      sizes="180x180"
+      href="/apple-icon-180x180.png"
+    />
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="192x192"
+      href="/android-icon-192x192.png"
+    />
+    <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
+    <link rel="icon" type="image/png" sizes="96x96" href="/favicon-96x96.png" />
+    <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
+    <meta name="msapplication-TileColor" content="#ffffff" />
+    <meta name="msapplication-TileImage" content="/ms-icon-144x144.png" />
+    <meta name="theme-color" content="#ffffff" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <!-- <meta name="description" content="Sauti Market Price Data API" /> -->
 
+    <!-- facebook open graph tags-->
+    <meta property="og:site_name" content="Sauti Africa" />
+    <meta property="og:title" content="Sauti Africa Market Price Data API" />
+    <meta
+      property="og:description"
+      content="A public-facing API that allows you to access Sauti Africa's market price data."
+    />
+    <meta property="og:type" content="website" />
+    <meta property="og:url" content="http://price-api.sautiafrica.org/" />
+    <meta
+      property="og:image"
+      content="https://images.unsplash.com/photo-1497528892195-b4ee7ea4f8bc?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80"
+    />
+    <link rel="apple-touch-icon" href="logo192.png" />
 
-  <meta charset="utf-8" />
-  <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <!-- <meta name="description" content="Sauti Market Price Data API" /> -->
+    <!-- twitter card tags -->
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:domain" value="http://price-api.sautiafrica.org" />
+    <meta name="twitter:title" value="Sauti Africa Market Price Data API" />
+    <meta
+      name="twitter:description"
+      value="A public-facing API that allows you to access Sauti Africa's market price data."
+    />
+    <meta
+      name="twitter:image"
+      content="https://images.unsplash.com/photo-1497528892195-b4ee7ea4f8bc?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80"
+    />
+    <meta name="twitter:url" value="http://price-api.sautiafrica.org" />
 
-  <!-- facebook open graph tags-->
-  <meta property="og:site_name" content="Sauti Africa" />
-  <meta property="og:title" content="Sauti Africa Market Price Data API" />
-  <meta property="og:description"
-    content="A public-facing API that allows you to access Sauti Africa's market price data." />
-  <meta property="og:type" content="website" />
-  <meta property="og:url" content="http://price-api.sautiafrica.org/" />
-  <meta property="og:image"
-    content="https://images.unsplash.com/photo-1497528892195-b4ee7ea4f8bc?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80" />
-  <link rel="apple-touch-icon" href="logo192.png" />
-
-  <!-- twitter card tags -->
-  <meta name="twitter:card" content="summary_large_image" />
-  <meta name="twitter:domain" value="http://price-api.sautiafrica.org" />
-  <meta name="twitter:title" value="Sauti Africa Market Price Data API" />
-  <meta name="twitter:description"
-    value="A public-facing API that allows you to access Sauti Africa's market price data." />
-  <meta name="twitter:image"
-    content="https://images.unsplash.com/photo-1497528892195-b4ee7ea4f8bc?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1350&q=80" />
-  <meta name="twitter:url" value="http://price-api.sautiafrica.org" />
-
-  <!--
+    <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
     -->
-  <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-  <!--
+    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
       Only files inside the `public` folder can be referenced from the HTML.
@@ -56,13 +105,13 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-  <title>Sauti Market Price Data API</title>
-</head>
+    <title>Sauti Market Price Data API</title>
+  </head>
 
-<body>
-  <noscript>You need to enable JavaScript to run this app.</noscript>
-  <div id="root"></div>
-  <!--
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+    <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
 
@@ -72,6 +121,5 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
-</body>
-
+  </body>
 </html>
